requirement.txt
Click        7.0    
coremltools  2.0    
Flask        1.0.2  
itsdangerous 1.1.0  
Jinja2       2.10   
MarkupSafe   1.1.0  
numpy        1.15.4 
Pillow       5.3.0  
pip          18.1   
protobuf     3.6.1  
scikit-learn 0.20.2 
scipy        1.2.0  
setuptools   40.6.3 
six          1.10.0 
sklearn      0.0    
Werkzeug     0.14.1 
wheel        0.32.3 

